+++ examples/lightspark-remote-signing-server/src/main.rs
@@ -1,6 +1,6 @@
 use actix_web::{get, post, web, App, HttpRequest, HttpResponse, HttpServer, Responder};
 use futures_util::StreamExt as _;
-use lightspark::{
+use lightspark_remote_signing::lightspark::{
     key::Secp256k1SigningKey, request::auth_provider::AccountAuthProvider, webhooks::WebhookEvent,
 };
 use lightspark_remote_signing::{
@@ -24,14 +24,20 @@ async fn webhook_handler(
     data: web::Data<config::Config>,
 ) -> impl Responder {
     let headers = req.headers();
-    let signature = headers.get(lightspark::webhooks::SIGNATURE_HEADER).unwrap();
+    let signature = headers
+        .get(lightspark_remote_signing::lightspark::webhooks::SIGNATURE_HEADER)
+        .unwrap();
     let mut bytes = web::BytesMut::new();
     while let Some(item) = body.next().await {
         bytes.extend_from_slice(&item.unwrap());
     }
 
     let auth = AccountAuthProvider::new(data.api_client_id.clone(), data.api_client_secret.clone());
-    let client = lightspark::client::LightsparkClient::<Secp256k1SigningKey>::new(auth).unwrap();
+    let client = lightspark_remote_signing::lightspark::client::LightsparkClient::<
+        Secp256k1SigningKey,
+    >::new(auth)
+    .unwrap();
+    // client.requester.set_base_url(data.api_endpoint.clone());
 
     let seed = Seed::new(hex::decode(data.master_seed_hex.clone()).unwrap());
     let signer =
@@ -63,7 +69,7 @@ async fn main() -> std::io::Result<()> {
             .service(ping)
             .service(webhook_handler)
     })
-    .bind(("127.0.0.1", 8080))?
+    .bind(("0.0.0.0", 8080))?
     .run()
     .await
 }