+++ lightspark/src/objects/lightspark_node_with_o_s_k.rs
@@ -1,27 +1,27 @@
 // Copyright ©, 2023-present, Lightspark Group, Inc. - All Rights Reserved
+use crate::error::Error;
 use crate::objects::bitcoin_network::BitcoinNetwork;
 use crate::objects::blockchain_balance::BlockchainBalance;
-use crate::objects::channel_status::ChannelStatus;
-use crate::objects::currency_amount::CurrencyAmount;
 use crate::objects::entity::Entity;
 use crate::objects::lightspark_node::LightsparkNode;
+use crate::objects::lightspark_node_status::LightsparkNodeStatus;
 use crate::objects::lightspark_node_to_channels_connection::LightsparkNodeToChannelsConnection;
+use crate::objects::node::Node;
 use crate::objects::node_address_type::NodeAddressType;
+use crate::objects::node_to_addresses_connection::NodeToAddressesConnection;
 use crate::objects::secret::Secret;
 use crate::types::custom_date_formats::custom_date_format;
 use crate::types::get_entity::GetEntity;
+use crate::types::graphql_requester::GraphQLRequester;
 use chrono::{DateTime, Utc};
 use serde::Deserialize;
+use serde_json::Value;
 use std::collections::HashMap;
 use std::vec::Vec;
 
-use crate::error::Error;
-use crate::objects::lightspark_node_status::LightsparkNodeStatus;
-use crate::objects::node::Node;
-use crate::objects::node_to_addresses_connection::NodeToAddressesConnection;
-use crate::request::requester::Requester;
+use crate::objects::channel_status::ChannelStatus;
+use crate::objects::currency_amount::CurrencyAmount;
 use crate::types::entity_wrapper::EntityWrapper;
-use serde_json::Value;
 
 /// This is a Lightspark node with OSK.
 #[derive(Clone, Deserialize)]
@@ -337,7 +337,7 @@ fragment LightsparkNodeWithOSKFragment on LightsparkNodeWithOSK {
 impl LightsparkNodeWithOSK {
     pub async fn get_addresses(
         &self,
-        requester: &Requester,
+        requester: &impl GraphQLRequester,
         first: Option<i64>,
         types: Option<Vec<NodeAddressType>>,
     ) -> Result<NodeToAddressesConnection, Error> {
@@ -362,10 +362,7 @@ impl LightsparkNodeWithOSK {
         variables.insert("types", types.into());
 
         let value = serde_json::to_value(variables).map_err(Error::ConversionError)?;
-        let result = requester
-            .execute_graphql(query, Some(value))
-            .await
-            .map_err(Error::ClientError)?;
+        let result = requester.execute_graphql(query, Some(value)).await?;
         let json = result["entity"]["addresses"].clone();
         let result = serde_json::from_value(json).map_err(Error::JsonError)?;
         Ok(result)
@@ -373,7 +370,7 @@ impl LightsparkNodeWithOSK {
 
     pub async fn get_channels(
         &self,
-        requester: &Requester,
+        requester: &impl GraphQLRequester,
         first: Option<i64>,
         statuses: Option<Vec<ChannelStatus>>,
         after: Option<String>,
@@ -496,10 +493,7 @@ impl LightsparkNodeWithOSK {
         variables.insert("after", after.into());
 
         let value = serde_json::to_value(variables).map_err(Error::ConversionError)?;
-        let result = requester
-            .execute_graphql(query, Some(value))
-            .await
-            .map_err(Error::ClientError)?;
+        let result = requester.execute_graphql(query, Some(value)).await?;
         let json = result["entity"]["channels"].clone();
         let result = serde_json::from_value(json).map_err(Error::JsonError)?;
         Ok(result)